feat(/form.js): Add relationship between form and reservations

+++ controllers/form.js
@@ -1,12 +1,16 @@
 import Form from '../models/solicitud.js'
+import Reservation from '../models/reservas.js'
+import Product from '../models/productos.js'
+import Teacher from '../models/docentes.js'
+import Student from '../models/estudiantes.js'
 import { handleError, handlePromiseError } from '../utils/error.js'
 
 export const getForms = async (_, res, next) => {
   try {
     const totalItems = await Form.find().countDocuments()
     const forms = await Form.find()
-      .populate('codigoDocente')
-      .populate('matriculaEstudiante')
+      .populate('idDocente')
+      .populate('idEstudiante')
       .sort({ createdAt: -1 })
 
     if (!forms) {
@@ -28,8 +32,12 @@ export const getForm = async (req, res, next) => {
 
   try {
     const form = await Form.findById(formId)
+      .populate('idDocente')
+      .populate('idEstudiante')
+      .sort({ createdAt: -1 })
+
     if (!form) {
-      throw Error(404, 'Could not find the form.')
+      throw handleError(404, 'Could not find the form.')
     }
 
     res.status(200).json({ message: 'Form Fetched', form })
@@ -48,6 +56,7 @@ export const postForm = async (req, res, next) => {
     telefono,
     matriculaEstudiante,
     codigoDocente,
+    equipos,
   } = req.body
 
   try {
@@ -59,9 +68,22 @@ export const postForm = async (req, res, next) => {
       !curso ||
       !telefono ||
       !matriculaEstudiante ||
-      !codigoDocente
+      !codigoDocente ||
+      !equipos
     ) {
       throw handleError(422, 'Validation Failed, data is incorrect')
+    }
+
+    // Check if student or teacher exists
+
+    const teacher = await Teacher.findOne({ codigo: codigoDocente })
+    if (!teacher) {
+      throw handleError(404, 'Invalid teacher')
+    }
+
+    const student = await Student.findOne({ matricula: matriculaEstudiante })
+    if (!student) {
+      throw handleError(404, 'Invalid student')
     }
 
     const form = new Form({
@@ -71,15 +93,37 @@ export const postForm = async (req, res, next) => {
       asignatura,
       curso,
       telefono,
-      codigoDocente,
-      matriculaEstudiante,
+      idDocente: teacher._id,
+      idEstudiante: student._id,
     })
 
+    // Create Revervation
+    const reservation = new Reservation()
+
+    for (const item of equipos) {
+      const product = await Product.findOne({ tipo: item, stock: { $gte: 1 } })
+
+      if (!product) {
+        throw handleError(404, `The ${item} is missing or out of stock`)
+      }
+
+      reservation.equipos.push(product)
+    }
+
+    reservation.idForm = form._id
+
     await form.save()
+    await reservation.save()
+
+    // Decrease Stock
+    for (const item of reservation.equipos) {
+      await Product.findByIdAndUpdate(item, { $inc: { stock: -1 } })
+    }
 
     res.status(201).json({
       message: 'Form has been created!',
       form,
+      reservation,
     })
   } catch (err) {
     handlePromiseError(err, next)
@@ -97,6 +141,7 @@ export const updateForm = async (req, res, next) => {
     telefono,
     codigoDocente,
     matriculaEstudiante,
+    equipos
   } = req.body
 
   try {
@@ -108,7 +153,8 @@ export const updateForm = async (req, res, next) => {
       !curso ||
       !telefono ||
       !matriculaEstudiante ||
-      !codigoDocente
+      !codigoDocente ||
+      !equipos
     ) {
       throw handleError(422, 'Validation Failed, data is incorrect')
     }
@@ -117,14 +163,27 @@ export const updateForm = async (req, res, next) => {
       throw handleError(404, 'Could not find post.')
     }
 
+    // Check if student or teacher exists
+
+    const teacher = await Teacher.findOne({ codigo: codigoDocente })
+    if (!teacher) {
+      throw handleError(404, 'Invalid teacher')
+    }
+
+    const student = await Student.findOne({ matricula: matriculaEstudiante })
+    if (!student) {
+      throw handleError(404, 'Invalid student')
+    }
+
     form.fechaDeUso = fechaDeUso
     form.horaInicio = horaInicio
     form.horaFinal = horaFinal
     form.asignatura = asignatura
     form.curso = curso
     form.telefono = telefono
-    form.codigoDocente = codigoDocente
-    form.matriculaEstudiante = matriculaEstudiante
+    form.idDocente = teacher._id
+    form.idEstudiante = student._id
+    form.equipos = equipos
 
     const savedForm = await form.save()
 
@@ -145,6 +204,16 @@ export const deleteForm = async (req, res, next) => {
       throw handleError(404, 'Could not find the form.')
     }
 
+    const reservation = await Reservation.findOne({ formId: formId })
+    if (!reservation) {
+      throw handleError(404, 'Could not find the reservation.')
+    }
+
+    for (const item of reservation.equipos) {
+      await Product.findByIdAndUpdate(item, { $inc: { stock: 1 } })
+    }
+
+    await Reservation.findOneAndRemove({ formId: formId })
     await Form.findByIdAndRemove(formId)
 
     res.status(200).json({ message: 'The form has been deleted.' })
